Move test guidance component to a shared spot
We might want to use this component elsewhere too so no longer keep in
under Schools
