+++ app/views/schools/register_ect/find_ect.html.erb
@@ -22,6 +22,6 @@
   <%= f.govuk_submit "Continue" %>
 <% end %>
 
-<%= render Schools::TestGuidanceComponent.new do |component| %>
+<%= render TestGuidanceComponent.new do |component| %>
   <% component.with_trs_example_teacher_details %>
 <% end %>
